requirements.txt
# API clients
openai>=1.12.0
anthropic>=0.16.0  # Keeping for backwards compatibility

# Framework requirements
crewai>=0.28.0
pytest>=7.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.2

# Optional LangChain dependencies (if needed)
langchain>=0.1.0
langchain-core>=0.1.0
